requirements.txt
aiofiles==0.4.0
asyncpg==0.24.0
dependency-injector==4.39.1
fastapi==0.78.0
fastapi-jwt-auth==0.5.0
fastapi-utils==0.2.1
Jinja2==3.1.2
pydantic==1.9.1
python-jose==3.3.0
requests==2.28.0
SQLAlchemy==1.4.25
starlette==0.19.1
uvicorn==0.17.6
